fixed reference to incompatible types.

+++ NeoVM/Types.h
@@ -2,16 +2,13 @@
 
 // Types
 
-typedef unsigned char byte;
-
-typedef short int16;
-typedef unsigned short uint16;
-
-typedef int int32;
-typedef unsigned int uint32;
-
-typedef long int64;
-typedef unsigned long uint64;
+typedef unsigned char           byte;
+typedef short                   int16;
+typedef unsigned short          uint16;
+typedef int                     int32;
+typedef unsigned int            uint32;
+typedef long long               int64;
+typedef unsigned long long      uint64;
 
 #ifndef NULL
 #define NULL 0
